+++ include/xemmai/dictionary.h
@@ -160,7 +160,7 @@ struct t_type_of<t_dictionary> : t_type
 {
 	static void f_construct(t_object* a_module, const t_value& a_self, t_slot* a_stack, size_t a_n);
 	static std::wstring f_string(const t_value& a_self);
-	static int f_hash(const t_value& a_self);
+	static ptrdiff_t f_hash(const t_value& a_self);
 	static bool f_equals(const t_value& a_self, const t_value& a_other);
 	static bool f_not_equals(const t_value& a_self, const t_value& a_other)
 	{
