+++ src/array.cc
@@ -54,13 +54,13 @@ void t_array::f_shrink()
 	}
 }
 
-void t_array::f_validate(int& a_index) const
+void t_array::f_validate(ptrdiff_t& a_index) const
 {
 	if (a_index < 0) {
 		a_index += v_size;
 		if (a_index < 0) t_throwable::f_throw(L"out of range.");
 	} else {
-		if (a_index >= static_cast<int>(v_size)) t_throwable::f_throw(L"out of range.");
+		if (a_index >= static_cast<ptrdiff_t>(v_size)) t_throwable::f_throw(L"out of range.");
 	}
 }
 
@@ -80,14 +80,14 @@ void t_array::f_swap(t_transfer& a_tuple, size_t& a_head, size_t& a_size)
 	std::swap(v_size, a_size);
 }
 
-const t_slot& t_array::operator[](int a_index) const
+const t_slot& t_array::operator[](ptrdiff_t a_index) const
 {
 	f_validate(a_index);
 	const t_tuple& tuple = f_as<const t_tuple&>(v_tuple);
 	return tuple[(v_head + a_index) % tuple.f_size()];
 }
 
-t_slot& t_array::operator[](int a_index)
+t_slot& t_array::operator[](ptrdiff_t a_index)
 {
 	f_validate(a_index);
 	t_tuple& tuple = f_as<t_tuple&>(v_tuple);
@@ -133,13 +133,13 @@ t_transfer t_array::f_shift()
 	return p;
 }
 
-void t_array::f_insert(int a_index, const t_transfer& a_value)
+void t_array::f_insert(ptrdiff_t a_index, const t_transfer& a_value)
 {
 	if (a_index < 0) {
 		a_index += v_size;
 		if (a_index < 0) t_throwable::f_throw(L"out of range.");
 	} else {
-		if (a_index > static_cast<int>(v_size)) t_throwable::f_throw(L"out of range.");
+		if (a_index > static_cast<ptrdiff_t>(v_size)) t_throwable::f_throw(L"out of range.");
 	}
 	f_grow();
 	t_tuple& tuple = f_as<t_tuple&>(v_tuple);
@@ -147,7 +147,7 @@ void t_array::f_insert(int a_index, const t_transfer& a_value)
 	size_t j = v_head + v_size;
 	size_t n = tuple.f_size();
 	t_slot* p = &tuple[0];
-	if (a_index < static_cast<int>(v_size / 2)) {
+	if (a_index < static_cast<ptrdiff_t>(v_size / 2)) {
 		if (i >= n) {
 			*f_move_backward(p + --v_head, p + n - 1) = p->f_transfer();
 			*f_move_backward(p, p + --i - n) = a_value;
@@ -173,7 +173,7 @@ void t_array::f_insert(int a_index, const t_transfer& a_value)
 	++v_size;
 }
 
-t_transfer t_array::f_remove(int a_index)
+t_transfer t_array::f_remove(ptrdiff_t a_index)
 {
 	f_validate(a_index);
 	t_tuple& tuple = f_as<t_tuple&>(v_tuple);
@@ -182,7 +182,7 @@ t_transfer t_array::f_remove(int a_index)
 	size_t n = tuple.f_size();
 	t_slot* p = &tuple[0];
 	t_transfer q = p[i % n].f_transfer();
-	if (a_index < static_cast<int>(v_size / 2)) {
+	if (a_index < static_cast<ptrdiff_t>(v_size / 2)) {
 		if (i >= n) {
 			*f_move_forward(p, p + i - n) = p[n - 1].f_transfer();
 			f_move_forward(p + v_head, p + n - 1);
@@ -244,11 +244,11 @@ std::wstring t_type_of<t_array>::f_string(const t_value& a_self)
 	return L'[' + s + L']';
 }
 
-int t_type_of<t_array>::f_hash(const t_value& a_self)
+ptrdiff_t t_type_of<t_array>::f_hash(const t_value& a_self)
 {
 	f_check<t_array>(a_self, L"this");
 	const t_array& array = f_as<const t_array&>(a_self);
-	int n = 0;
+	ptrdiff_t n = 0;
 	size_t i = 0;
 	while (true) {
 		t_transfer x;
@@ -258,8 +258,8 @@ int t_type_of<t_array>::f_hash(const t_value& a_self)
 			x = array[i];
 		}
 		x = x.f_hash();
-		f_check<int>(x, L"value");
-		n ^= f_as<int>(x);
+		f_check<ptrdiff_t>(x, L"value");
+		n ^= f_as<ptrdiff_t>(x);
 		++i;
 	}
 	return n;
@@ -467,9 +467,9 @@ void t_type_of<t_array>::f_define()
 	t_define<t_array, t_object>(f_global(), L"Array")
 		(f_global()->f_symbol_construct(), f_construct)
 		(f_global()->f_symbol_string(), t_member<std::wstring (*)(const t_value&), f_string>())
-		(f_global()->f_symbol_hash(), t_member<int (*)(const t_value&), f_hash>())
-		(f_global()->f_symbol_get_at(), t_member<const t_value& (t_array::*)(int) const, &t_array::f_get_at, t_with_lock_for_read>())
-		(f_global()->f_symbol_set_at(), t_member<const t_value& (t_array::*)(int, const t_transfer&), &t_array::f_set_at, t_with_lock_for_write>())
+		(f_global()->f_symbol_hash(), t_member<ptrdiff_t (*)(const t_value&), f_hash>())
+		(f_global()->f_symbol_get_at(), t_member<const t_value& (t_array::*)(ptrdiff_t) const, &t_array::f_get_at, t_with_lock_for_read>())
+		(f_global()->f_symbol_set_at(), t_member<const t_value& (t_array::*)(ptrdiff_t, const t_transfer&), &t_array::f_set_at, t_with_lock_for_write>())
 		(f_global()->f_symbol_less(), t_member<bool (*)(const t_value&, const t_value&), f_less>())
 		(f_global()->f_symbol_less_equal(), t_member<bool (*)(const t_value&, const t_value&), f_less_equal>())
 		(f_global()->f_symbol_greater(), t_member<bool (*)(const t_value&, const t_value&), f_greater>())
@@ -482,8 +482,8 @@ void t_type_of<t_array>::f_define()
 		(L"pop", t_member<t_transfer (t_array::*)(), &t_array::f_pop, t_with_lock_for_write>())
 		(L"unshift", t_member<void (t_array::*)(const t_transfer&), &t_array::f_unshift, t_with_lock_for_write>())
 		(L"shift", t_member<t_transfer (t_array::*)(), &t_array::f_shift, t_with_lock_for_write>())
-		(L"insert", t_member<void (t_array::*)(int, const t_transfer&), &t_array::f_insert, t_with_lock_for_write>())
-		(L"remove", t_member<t_transfer (t_array::*)(int), &t_array::f_remove, t_with_lock_for_write>())
+		(L"insert", t_member<void (t_array::*)(ptrdiff_t, const t_transfer&), &t_array::f_insert, t_with_lock_for_write>())
+		(L"remove", t_member<t_transfer (t_array::*)(ptrdiff_t), &t_array::f_remove, t_with_lock_for_write>())
 		(L"each", t_member<void (*)(const t_value&, const t_value&), f_each>())
 		(L"sort", t_member<void (*)(const t_value&, const t_value&), f_sort>())
 	;
@@ -524,9 +524,9 @@ void t_type_of<t_array>::f_get_at(t_object* a_this, t_slot* a_stack)
 {
 	t_native_context context;
 	t_transfer a0 = a_stack[1].f_transfer();
-	f_check<int>(a0, L"index");
+	f_check<ptrdiff_t>(a0, L"index");
 	t_with_lock_for_read lock(a_this);
-	a_stack[0].f_construct(f_as<const t_array&>(a_this).f_get_at(f_as<int>(a0)));
+	a_stack[0].f_construct(f_as<const t_array&>(a_this).f_get_at(f_as<ptrdiff_t>(a0)));
 	context.f_done();
 }
 
@@ -535,9 +535,9 @@ void t_type_of<t_array>::f_set_at(t_object* a_this, t_slot* a_stack)
 	t_native_context context;
 	t_transfer a0 = a_stack[1].f_transfer();
 	t_transfer a1 = a_stack[2].f_transfer();
-	f_check<int>(a0, L"index");
+	f_check<ptrdiff_t>(a0, L"index");
 	t_with_lock_for_write lock(a_this);
-	a_stack[0].f_construct(f_as<t_array&>(a_this).f_set_at(f_as<int>(a0), a1));
+	a_stack[0].f_construct(f_as<t_array&>(a_this).f_set_at(f_as<ptrdiff_t>(a0), a1));
 	context.f_done();
 }
